+++ branches/ssg/sdl/common/lsys.c
@@ -519,8 +519,8 @@ static void lsysi_dobang(struct lsys_turtlestatei *cmd)
 
 static void lsysi_dosizedm(struct lsys_turtlestatei *cmd)
 {
-  LDBL angle = (LDBL) cmd->realangle * ANGLE2DOUBLE;
-  LDBL s, c;
+  double angle = (double) cmd->realangle * ANGLE2DOUBLE;
+  double s, c;
   long fixedsin, fixedcos;
 
   FPUsincos(&angle, &s, &c);
@@ -552,8 +552,8 @@ static void lsysi_dosizegf(struct lsys_turtlestatei *cmd)
 
 static void lsysi_dodrawd(struct lsys_turtlestatei *cmd)
 {
-  LDBL angle = (LDBL) cmd->realangle * ANGLE2DOUBLE;
-  LDBL s, c;
+  double angle = (double) cmd->realangle * ANGLE2DOUBLE;
+  double s, c;
   long fixedsin, fixedcos;
   int lastx, lasty;
 
@@ -572,8 +572,8 @@ static void lsysi_dodrawd(struct lsys_turtlestatei *cmd)
 
 static void lsysi_dodrawm(struct lsys_turtlestatei *cmd)
 {
-  LDBL angle = (LDBL) cmd->realangle * ANGLE2DOUBLE;
-  LDBL s, c;
+  double angle = (double) cmd->realangle * ANGLE2DOUBLE;
+  double s, c;
   long fixedsin, fixedcos;
 
   FPUsincos(&angle, &s, &c);
@@ -1069,11 +1069,11 @@ static struct lsys_cmd *
 
 static void lsysi_dosincos(void)
 {
-  LDBL locaspect;
-  LDBL TWOPI = 2.0 * PI;
-  LDBL twopimax;
-  LDBL twopimaxi;
-  LDBL s, c;
+  double locaspect;
+  double TWOPI = 2.0 * PI;
+  double twopimax;
+  double twopimaxi;
+  double s, c;
   int i;
 
   locaspect=screenaspect*xdots/ydots;
